Added transaction services. Added pagination support. Fixed test.

+++ transactions-core/src/main/java/demo/services/AccountService.java
@@ -1,27 +1,32 @@
 package demo.services;
 
-import java.util.List;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import demo.model.Account;
 import demo.repo.AccountRepository;
+import demo.repo.TransactionRepository;
 
 @Service
 @Transactional
 public class AccountService {
 
     private static final Logger logger = LoggerFactory.getLogger(AccountService.class);
-    
+
     @Autowired
     private AccountRepository accountRepository;
 
-    public List<Account> list() {
-        return accountRepository.findAll();
+    @Autowired
+    private TransactionRepository transactionRepository;
+
+    public Page<Account> list(Pageable pageable) {
+        return accountRepository.findAll(pageable);
     }
 
     public Account findOne(Long id) {
@@ -30,22 +35,28 @@ public class AccountService {
 
     public Account save(Account account) {
         final Account saved = accountRepository.save(account);
-        
+
         logger.info("Created {}", saved);
         return saved;
     }
 
     public Account update(Account account) {
         final Account saved = accountRepository.save(account);
-        
+
         logger.info("Updated {}", saved);
-        
+
         return saved;
     }
 
     public void delete(Long id) {
+        boolean noTrasactions = transactionRepository.findByAccountId(id, new PageRequest(0, 1)).getContent().isEmpty();
+
+        if (!noTrasactions) {
+            throw new IllegalArgumentException("Account (id: " + id + ") cannot be deleted since it has transactions.");
+        }
+
         accountRepository.delete(id);
-        
+
         logger.info("Deleted account {}", id);
     }
